review_list_page/api/pyproject.toml
[tool.poetry]
name = "review_list_page"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
indico-client = "^4.5.0"
uvicorn = "^0.14.0"
uvloop = "^0.15.2"
httptools = "^0.2.0"
gunicorn = "^20.1.0"
virga = {git = "https://github.com/IndicoDataSolutions/virga.git", rev = "main"}
asyncpg = "^0.24.0"
aiodataloader = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
watchgod = "^0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
